requirements.txt
jsonlines
nicegui
pytest-asyncio
pytest-selenium
pytest
pytest-order